Get FavIcon from WebView

+++ qml/js/Favorites.js
@@ -1,23 +1,23 @@
 .pragma library
 
 var defaultfavorites = [ { "title": "Jolla", "url": "https://jolla.com/" },
-                         { "title": "Jolla Together", "url": "https://together.jolla.com/" },
+                         { "title": "Jolla Together", "url": "https://together.jolla.com/questions/" },
                          { "title": "Jolla Users", "url": "http://www.jollausers.com/" },
                          { "title": "Maemo Talk", "url": "http://talk.maemo.org/" },
-                         { "title": "Sailfish OS", "url": "http://www.sailfishos.org/"} ];
+                         { "title": "Sailfish OS", "url": "https://www.sailfishos.org/"} ];
 
 function load(db, favorites)
 {
     var i = 0;
 
     db.transaction(function(tx) {
-        tx.executeSql("CREATE TABLE IF NOT EXISTS Favorites (url TEXT PRIMARY KEY, title TEXT)");
+        tx.executeSql("CREATE TABLE IF NOT EXISTS Favorites (url TEXT PRIMARY KEY, title TEXT, icon TEXT)");
         var res = tx.executeSql("SELECT * FROM Favorites;");
 
         if(res.rows.length > 0)
         {
             for(i = 0; i < res.rows.length; i++)
-                favorites.append({ "title": res.rows[i].title, "url": res.rows[i].url });
+                favorites.append({ "title": res.rows[i].title, "url": res.rows[i].url, "icon": res.rows[i].icon });
         }
         else
         {
@@ -30,36 +30,50 @@ function load(db, favorites)
 
 function add(db, favorites, title, url)
 {
-    favorites.append({ "title": title, "url": url });
+    favorites.append({ "title": title, "url": url, "icon": "" });
 
     db.transaction(function(tx) {
-        tx.executeSql("INSERT OR REPLACE INTO Favorites (url, title) VALUES (?, ?);", [url, title]);
+        tx.executeSql("INSERT OR REPLACE INTO Favorites (url, title, icon) VALUES (?, ?, ?);", [url, title, ""]);
     });
 }
 
 function replace(db, favorites, index, title, url)
 {
     db.transaction(function(tx) {
-        tx.executeSql("UPDATE Favorites SET url=?, title=? WHERE url=? AND title=?;", [url, title, favorites.get(index).url, favorites.get(index).title]);
+        tx.executeSql("UPDATE Favorites SET url=?, title=?, icon='' WHERE url=? AND title=?;", [url, title, favorites.get(index).url, favorites.get(index).title]);
     });
 
-    var favorite = favorites.get(index)
-    favorite.url = url
-    favorite.title = title
+    var favorite = favorites.get(index);
+    favorite.url = url;
+    favorite.title = title;
+    favorite.icon = "";
 }
 
 function remove(db, favorites, url)
 {
-    for(var i = 0; i < favorites.count; i++)
-    {
-        if(favorites.get(i).url === url)
-        {
-            favorites.remove(i);
-            break;
-        }
-    }
+    var idx = favorites.indexOf(url);
+
+    if(idx === -1)
+        return;
+
+    favorites.remove(idx);
 
     db.transaction(function(tx) {
         tx.executeSql("DELETE FROM Favorites WHERE url= ?;", [url]);
     });
+}
+
+function setIcon(db, favorites, url, icon)
+{
+    var idx = favorites.indexOf(url);
+
+    if(idx === -1)
+        return;
+
+    db.transaction(function(tx) {
+        tx.executeSql("UPDATE Favorites SET icon=? WHERE url=?;", [icon, url]);
+    });
+
+    var favorite = favorites.get(idx);
+    favorite.icon = icon;
 }
